+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-* Add Django 5.0+ fixer to rewrite ``format_html()`` calls without ``args`` or ``kwargs`` probably using ``str.format()`` incorrectly.
+* Add Django 5.0+ fixer to rewrite ``format_html()`` calls without any arguments.
 
   `Issue #477 <https://github.com/adamchainz/django-upgrade/issues/477>`__.