+++ CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog
 =========
 
+1.21.0 (2024-09-05)
+-------------------
+
 * Add Django 5.0+ fixer to rewrite ``format_html()`` calls without ``args`` or ``kwargs`` probably using ``str.format()`` incorrectly.
 
   `Issue #477 <https://github.com/adamchainz/django-upgrade/issues/477>`__.